end first shot contact; move contactsheet to utils
